+++ app/event/index.js
@@ -1,11 +1,11 @@
 const raiseEvent = require('./raise-event')
 const sendSessionEvent = require('./send-session-event')
 const sendMonitoringEvent = require('./send-monitoring-event')
-const sendExceptionEvent = require('./send-exception-event')
+const raiseIneligibilityEvent = require('./raise-ineligibility-event')
 
 module.exports = {
   raiseEvent,
   sendSessionEvent,
   sendMonitoringEvent,
-  sendExceptionEvent
+  raiseIneligibilityEvent
 }
