+++ test/integration/narrow/routes/auth/signin-oidc.test.js
@@ -9,7 +9,7 @@ const personMock = require('../../../../../app/api-requests/rpa-api/person')
 jest.mock('../../../../../app/api-requests/rpa-api/person')
 const organisationMock = require('../../../../../app/api-requests/rpa-api/organisation')
 jest.mock('../../../../../app/api-requests/rpa-api/organisation')
-const sendExceptionEventMock = require('../../../../../app/event/send-exception-event')
+const sendExceptionEventMock = require('../../../../../app/event/raise-ineligibility-event')
 jest.mock('../../../../../app/event/send-exception-event')
 const cphCheckMock = require('../../../../../app/api-requests/rpa-api/cph-check')
 jest.mock('../../../../../app/api-requests/rpa-api/cph-check')
